Added file changing & manipulation to CS+ profiles
File support is no longer natively in IProfile, now implemented per case
Removed milliseconds from modify date
Corrected Y position for "set modify date to current date" button
Fixed OOB modify date being possible (1/1/1970 02:00AM is 0 in Unix
time, any earlier date would be negative)
Moved RadioBox into RadioBoxes class

+++ src/main/java/com/leo/cse/frontend/FrontUtils.java
@@ -135,7 +135,7 @@ public class FrontUtils {
 
 	public static void drawNineSlice(Graphics g, Image img, int x, int y, int w, int h) {
 		final int iw = img.getWidth(null), ih = img.getHeight(null);
-		final int sw = iw / 3, sh = ih / 3;
+		final int sw = (int) (iw / 3f), sh = (int) (ih / 3f);
 		// top left
 		g.drawImage(img, x, y, x + sw, y + sh, 0, 0, sw, sh, null);
 		// top middle
